New Polymesh sample notebook, new raster/vector output function, new function for diverging colormaps in bokeh
